Limit pitch based on edge insets. Fix max Z calculation in getProjMatrix.
As we don't want to show horizon, there is a need to limit max pitch based on edge insets.
The limit shouldn't be that bad, as to e.g. have 60 deg pitch, top - bottom insets should be less than 0.732 of screen height.

TransformState::getProjMatrix calculation of farZ was an aproximation. Replacing it with simpler, but precise calculation.

Related to: #15163